Move resource classes to resource package, add filter for form updates
