+++ src/main/java/org/eclipsefoundation/react/resources/OrganizationResource.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipsefoundation.react.request;
+package org.eclipsefoundation.react.resources;
 
 import java.util.ArrayList;
 import java.util.List;
