+++ src/main/java/org/eclipsefoundation/react/resources/WorkingGroupsResource.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipsefoundation.react.request;
+package org.eclipsefoundation.react.resources;
 
 import java.util.Arrays;
 import java.util.List;
@@ -30,7 +30,8 @@ import javax.ws.rs.core.Response;
 import org.eclipsefoundation.core.helper.CSRFHelper;
 import org.eclipsefoundation.core.namespace.DefaultUrlParameterNames;
 import org.eclipsefoundation.persistence.model.RDBMSQuery;
-import org.eclipsefoundation.react.dto.FormOrganization;
+import org.eclipsefoundation.react.dto.Contact;
+import org.eclipsefoundation.react.dto.FormWorkingGroup;
 import org.eclipsefoundation.react.dto.MembershipForm;
 import org.eclipsefoundation.react.namespace.MembershipFormAPIParameterNames;
 import org.jboss.resteasy.specimpl.MultivaluedMapImpl;
@@ -43,13 +44,13 @@ import io.quarkus.security.Authenticated;
  * @author Martin Lowe
  */
 @Authenticated
-@Path("form/{id}/organizations")
+@Path("form/{id}/working_groups")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class FormOrganizationsResource extends AbstractRESTResource {
+public class WorkingGroupsResource extends AbstractRESTResource {
 
     @GET
-    public Response getAll(@PathParam("id") String formID,
+    public Response getWorkingGroups(@PathParam("id") String formID,
             @HeaderParam(value = CSRFHelper.CSRF_HEADER_NAME) String csrf) {
         // ensure csrf
         csrfHelper.compareCSRF(aud, csrf);
@@ -58,12 +59,11 @@ public class FormOrganizationsResource extends AbstractRESTResource {
         if (r != null) {
             return r;
         }
-
         // create parameter map
         MultivaluedMap<String, String> params = new MultivaluedMapImpl<>();
         params.add(MembershipFormAPIParameterNames.FORM_ID.getName(), formID);
-        // retrieve the possible cached object
-        List<FormOrganization> results = dao.get(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class), params));
+        // retrieve the possible object
+        List<FormWorkingGroup> results = dao.get(new RDBMSQuery<>(wrap, filters.get(FormWorkingGroup.class), params));
         if (results == null) {
             return Response.serverError().build();
         }
@@ -72,89 +72,94 @@ public class FormOrganizationsResource extends AbstractRESTResource {
     }
 
     @POST
-    public Response create(@PathParam("id") String formID, FormOrganization org) {
+    public Response createWorkingGroup(@PathParam("id") String formID, FormWorkingGroup wg) {
         // check if user is allowed to modify these resources
         Response r = checkAccess(formID);
         if (r != null) {
             return r;
         }
-        // check if an org for this form already exists. If so, replace it with this one
-        MultivaluedMap<String, String> params = new MultivaluedMapImpl<>();
-        params.add(MembershipFormAPIParameterNames.FORM_ID.getName(), formID);
-        List<FormOrganization> results = dao.get(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class), params));
-        if (results != null && !results.isEmpty()) {
-            return update(formID, results.get(0).getId(), org);
-        } else {
-            MembershipForm form = dao.getReference(formID, MembershipForm.class);
-            org.setForm(form);
-            if (org.getAddress() != null) {
-                org.getAddress().setOrganization(org);
+        MembershipForm form = dao.getReference(formID, MembershipForm.class);
+        wg.setForm(dao.getReference(formID, MembershipForm.class));
+        // update the nested contact
+        if (wg.getContact() != null) {
+            if (wg.getContact().getId() != null) {
+                // update the contact object to get entity wg if set
+                Contact c = dao.getReference(wg.getContact().getId(), Contact.class);
+                wg.setContact(wg.getContact().cloneTo(c));
             }
+            // set the form back for wgerences
+            wg.getContact().setForm(form);
         }
-        return Response.ok(dao.add(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class)), Arrays.asList(org)))
+        return Response.ok(dao.add(new RDBMSQuery<>(wrap, filters.get(FormWorkingGroup.class)), Arrays.asList(wg)))
                 .build();
     }
 
     @GET
-    @Path("{orgID}")
-    public Response get(@PathParam("id") String formID, @PathParam("orgID") String id,
+    @Path("{wgID}")
+    public Response getWorkingGroup(@PathParam("id") String formID, @PathParam("wgID") String wgID,
             @HeaderParam(value = CSRFHelper.CSRF_HEADER_NAME) String csrf) {
         // ensure csrf
         csrfHelper.compareCSRF(aud, csrf);
+        // check if user is allowed to modify these resources
         Response r = checkAccess(formID);
         if (r != null) {
             return r;
         }
-
         // create parameter map
         MultivaluedMap<String, String> params = new MultivaluedMapImpl<>();
-        params.add(DefaultUrlParameterNames.ID.getName(), id);
+        params.add(DefaultUrlParameterNames.ID.getName(), wgID);
         params.add(MembershipFormAPIParameterNames.FORM_ID.getName(), formID);
-
-        // retrieve the possible cached object
-        List<FormOrganization> results = dao.get(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class), params));
+        // retrieve the possible object
+        List<FormWorkingGroup> results = dao.get(new RDBMSQuery<>(wrap, filters.get(FormWorkingGroup.class), params));
         if (results == null) {
             return Response.serverError().build();
         } else if (results.isEmpty()) {
             return Response.status(404).build();
         }
+        
         // return the results as a response
         return Response.ok(results.get(0)).build();
     }
 
     @PUT
-    @Path("{orgID}")
-    public Response update(@PathParam("id") String formID, @PathParam("orgID") String id, FormOrganization org) {
-        // need to fetch ref to use attached entity
+    @Path("{wgID}")
+    public Response updateWorkingGroup(@PathParam("id") String formID, FormWorkingGroup wg,
+            @PathParam("wgID") String wgID) {
+        // check if user is allowed to modify these resources
         Response r = checkAccess(formID);
         if (r != null) {
             return r;
         }
-        FormOrganization ref = dao.getReference(id, FormOrganization.class);
-        org.cloneTo(ref);
-        if (ref.getAddress() != null && org.getAddress() != null) {
-            // if both remote + new set, update remote values
-            org.getAddress().cloneTo(ref.getAddress());
+        // need to fetch ref to use attached entity
+        FormWorkingGroup ref = wg.cloneTo(dao.getReference(wgID, FormWorkingGroup.class));
+        ref.setForm(dao.getReference(formID, MembershipForm.class));
+        // update the nested contact
+        if (ref.getContact() != null && wg.getContact() != null) {
+            wg.getContact().cloneTo(ref.getContact());
         } else {
-            // if remote or new are missing, replace remote w/ new value
-            ref.setAddress(org.getAddress());
+            ref.setContact(wg.getContact());
+            // set the form back for references
+            if (ref.getContact() != null) {
+                ref.getContact().setForm(ref.getForm());
+            }
         }
-        return Response.ok(dao.add(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class)), Arrays.asList(ref)))
+        return Response.ok(dao.add(new RDBMSQuery<>(wrap, filters.get(FormWorkingGroup.class)), Arrays.asList(ref)))
                 .build();
     }
 
     @DELETE
-    @Path("{orgID}")
-    public Response delete(@PathParam("id") String formID, @PathParam("orgID") String id) {
+    @Path("{wgID}")
+    public Response deleteWorkingGroup(@PathParam("id") String formID, @PathParam("wgID") String id) {
+        // check if user is allowed to modify these resources
         Response r = checkAccess(formID);
         if (r != null) {
             return r;
         }
         MultivaluedMap<String, String> params = new MultivaluedMapImpl<>();
         params.add(DefaultUrlParameterNames.ID.getName(), id);
-        params.add(MembershipFormAPIParameterNames.FORM_ID.getName(), formID);
+        params.add(MembershipFormAPIParameterNames.USER_ID.getName(), ident.getPrincipal().getName());
 
-        dao.delete(new RDBMSQuery<>(wrap, filters.get(FormOrganization.class), params));
+        dao.delete(new RDBMSQuery<>(wrap, filters.get(FormWorkingGroup.class), params));
         return Response.ok().build();
     }
 }
